worker-service/requirements.txt
google-cloud-pubsub==2.21.1
python-dotenv==1.0.1
